pyproject.toml
[project]
name = "strawberry-spoke"
version = "0.1.0"
description = "Voice assistant spoke for Strawberry AI"
requires-python = ">=3.10"
readme = "README.md"

dependencies = [
    # Core
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    
    # Audio
    "sounddevice>=0.4",
    "numpy>=1.24",
]

[project.optional-dependencies]
# Voice processing (install separately due to size/licensing)
picovoice = [
    "pvporcupine>=3.0",
    "pvorca>=1.0",
    "pvcobra>=2.0",
    "pvleopard>=2.0",
]
google = [
    "google-cloud-speech>=2.0",
    "google-cloud-texttospeech>=2.0",
]
silero = [
    "torch>=2.0",
]
ui = [
    "PySide6>=6.5",
    "qasync>=0.24",
]
hub = [
    "tensorzero>=0.1.0",
]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "ruff>=0.1",
]
all = [
    "strawberry-spoke[picovoice,google,silero,ui,hub,dev]",
]

[project.scripts]
strawberry = "strawberry.main:main"
strawberry-test = "strawberry.testing.runner:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

